dedicated class to manage cozy nest config
