+++ scripts/CozyLogger.py
@@ -2,10 +2,10 @@
 # This is a simple logger class that will be used to log messages stdout
 # check the config if log is enabled
 class CozyLoggerClass:
-    def __init__(self, name: str, config):
+    def __init__(self, name: str, log_enabled):
 
         self.name = name
-        self.log_enabled = config["log_enabled"]
+        self.log_enabled = log_enabled
 
         if self.log_enabled:
             print(f"[{self.name}:DEBUG] Logger enabled. Disable in nevyui_settings.json")
