+++ scripts/tools.py
@@ -4,6 +4,28 @@ import os
 
 from PIL import Image
 
+from modules import shared, scripts
+
+
+def output_folder_array():
+    outdir_txt2img_samples = shared.opts.data['outdir_txt2img_samples']
+    outdir_img2img_samples = shared.opts.data['outdir_img2img_samples']
+    outdir_extras_samples = shared.opts.data['outdir_extras_samples']
+    base_dir = scripts.basedir()
+    # check if outdir_txt2img_samples is a relative path
+    if not os.path.isabs(outdir_txt2img_samples):
+        outdir_txt2img_samples = os.path.normpath(os.path.join(base_dir, outdir_txt2img_samples))
+    if not os.path.isabs(outdir_img2img_samples):
+        outdir_img2img_samples = os.path.normpath(os.path.join(base_dir, outdir_img2img_samples))
+    if not os.path.isabs(outdir_extras_samples):
+        outdir_extras_samples = os.path.normpath(os.path.join(base_dir, outdir_extras_samples))
+    images_folders = [
+        outdir_txt2img_samples,
+        outdir_img2img_samples,
+        outdir_extras_samples,
+    ]
+    return images_folders
+
 
 def get_image_exif(path: str):
     try:
